Changed images and masks of the loader precision to standard float32

+++ scripts/experiments/trainer.py
@@ -15,8 +15,8 @@ class Trainer:
         self.model.train()
         total_loss = 0
         for images, masks, _ in train_loader:
-            images = images.to(self.device, dtype=torch.float16)
-            masks  = masks.to(self.device, dtype=torch.float16)
+            images = images.to(self.device)
+            masks  = masks.to(self.device)
             self.optimizer.zero_grad()
             # Mixed precision forward pass
             with autocast('cuda'):
@@ -36,8 +36,8 @@ class Trainer:
         threshold = 0.5
         with torch.no_grad():
             for images, masks, paths in loader:
-                images = images.to(self.device, dtype=torch.float16)
-                masks = masks.to(self.device, dtype=torch.float16)
+                images = images.to(self.device)
+                masks = masks.to(self.device)
                 with autocast('cuda'):
                     outputs = self.model(images)
                     loss = self.criterion(outputs, masks)
